Single-Pass Training, Top-n Accuracy, and encoding_checker_utils
Added option for single-pass training, which splits the dataset into subsets and trains on each subset intensely before discarding it and using a new subset. The code maintains the same amount of training epochs and iterations.

Added top-n accuracy measure for measuring performance of the trained encoder.

Augmented the functions in encoding_checker_utils.py to have more functions and have a cleaner flow.